HeaderForLoggedUserElement was added.
