+++ kaliazinProject1/src/main/java/org/pages/HomePage.java
@@ -1,35 +1,21 @@
 package org.pages;
 
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
+import org.pages.elements.HeaderForLoggedUserElement;
 
 public class HomePage extends ParentPage{
-//    Logger logger = Logger.getLogger(HomePage.class.getName());
-
-    @FindBy(xpath = "//button[text()='Sign Out']")
-    private WebElement buttonSignOut;
-
-    @FindBy (xpath = "//a[@class='btn btn-sm btn-success mr-2']")
-    private WebElement buttonCreatePost;
 
     public HomePage(WebDriver webDriver) {
         super(webDriver);
     }
 
-    public void checkButtonSignOutVisible() {
-//        Assert.assertTrue("Button Sign Out is not visible", isButtonSignOutVisible());
-        checkIsElementDisplayed(buttonSignOut);
+    public HeaderForLoggedUserElement getHeaderForLoggedUserElement() {
+        return new HeaderForLoggedUserElement(webDriver);
     }
 
     public HomePage checkIsRedirectToHomePage() {
         //TODO check URL
-        checkButtonSignOutVisible();
+        getHeaderForLoggedUserElement().checkButtonSignOutVisible();
         return this;
-    }
-
-    public CreateNewPostPage clickOnButtonCreatePost() {
-        clickOnElement(buttonCreatePost);
-        return new CreateNewPostPage(webDriver);
     }
 }
