+++ kaliazinProject1/src/main/java/org/pages/PageProvider.java
@@ -16,4 +16,7 @@ public class PageProvider {
         return new HomePage(webDriver);
         }
 
+    public PostPage getPostPage() {
+        return new PostPage(webDriver);
+    }
 }
