+++ kaliazinProject1/src/main/java/org/pages/elements/HeaderForLoggedUserElement.java
@@ -1,0 +1,37 @@
+package org.pages.elements;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.pages.CommonActionsWithElements;
+import org.pages.CreateNewPostPage;
+
+public class HeaderForLoggedUserElement extends CommonActionsWithElements {
+
+    @FindBy(xpath = "//button[text()='Sign Out']")
+    private WebElement buttonSignOut;
+
+    @FindBy (xpath = "//a[@class='btn btn-sm btn-success mr-2']")
+    private WebElement buttonCreatePost;
+
+    @FindBy (xpath = "//img[@alt='My profile']")
+    private WebElement buttonMyProfile;
+
+    public HeaderForLoggedUserElement(WebDriver webDriver) {
+        super(webDriver);
+    }
+    public void clickOnButtonMyProfile() {
+        clickOnElement(buttonMyProfile);
+    }
+    public void checkButtonSignOutVisible() {
+        checkIsElementDisplayed(buttonSignOut);
+    }
+
+public CreateNewPostPage clickOnButtonCreatePost() {
+    clickOnElement(buttonCreatePost);
+    return new CreateNewPostPage(webDriver);
+ }
+}
+
+
+
